+++ docs/source/conf.py
@@ -65,6 +65,7 @@ numfig_format = {
 intersphinx_mapping = {
     'numpy': ('https://numpy.org/doc/stable/', None),
     'scipy': ('https://docs.scipy.org/doc/scipy/', None),
+    'matplotlib': ('https://matplotlib.org/stable/', None),
     }
 
 mathjax3_config = {